+++ e2e_tests/tests/cli.test.js
@@ -16,9 +16,7 @@ describe('KYT CLI', () => {
       expect(shell.test('-d', 'src')).toBe(true);
       expect(shell.test('-f', 'kyt.config.js')).toBe(true);
       expect(shell.test('-f', '.editorconfig')).toBe(true);
-      expect(shell.test('-f', '.babelrc')).toBe(true);
-      expect(shell.test('-f', '.eslintrc.json')).toBe(true);
-      expect(shell.test('-f', '.stylelintrc.json')).toBe(true);
+      expect(shell.test('-f', '.eslintrc.js')).toBe(true);
       expect(shell.test('-f', 'prototype.js')).toBe(true);
     });
   };
@@ -31,6 +29,8 @@ describe('KYT CLI', () => {
       expect(scripts.dev).toBe('kyt dev');
       expect(scripts.start).toBe('node build/server/main.js');
       expect(scripts.build).toBe('kyt build');
+      expect(scripts.test).toBe('jest');
+      expect(scripts.lint).toBe('eslint .');
       expect(scripts.proto).toBe('kyt proto');
       expect(scripts['kyt:help']).toBe('kyt --help');
     });
@@ -40,11 +40,9 @@ describe('KYT CLI', () => {
     expect(setupArr.includes('👍  Added kyt scripts into your package.json scripts')).toBe(true);
     expect(setupArr.includes('👍  Added new dependencies to package.json')).toBe(true);
     expect(setupArr.includes('👍  Installed new modules')).toBe(true);
-    expect(setupArr.includes('👍  Created .eslintrc.json file')).toBe(true);
-    expect(setupArr.includes('👍  Created .stylelintrc.json file')).toBe(true);
+    expect(setupArr.includes('👍  Created .eslintrc.js file')).toBe(true);
     expect(setupArr.includes('👍  Created kyt.config.js file')).toBe(true);
     expect(setupArr.includes('👍  Created .editorconfig file')).toBe(true);
-    expect(setupArr.includes('👍  Created .babelrc')).toBe(true);
     expect(setupArr.includes('👍  Created .gitignore file')).toBe(true);
     expect(setupArr.includes('👍  Created src directory')).toBe(true);
   };