+++ packages/kyt-starter-universal/starter-src/jest.config.js
@@ -1,3 +1,14 @@
+const path = require('path');
+
 module.exports = {
-  preset: 'jest-preset-kyt',
+  projects: [
+    require.resolve('./test/jest.test.config'),
+    require.resolve('./test/jest.style.config'),
+  ],
+  cacheDirectory: path.resolve(__dirname, '.caches/jest'),
+  haste: {
+    computeSha1: true,
+    providesModuleNodeModules: [],
+    throwOnModuleCollision: false,
+  },
 };